IFetchDataForViewItemResult moved to types
