+++ SharePointFramework/PortfolioWebParts/src/components/PortfolioAggregation/itemColumn.tsx
@@ -89,7 +89,11 @@ export const renderItemColumn = (item: any, index: number, column: IColumn) => {
             })}
             styles={{ root: { verticalAlign: 'bottom' } }}
           />
-          <Link href={item.ServerRedirectedURL} rel='noopener noreferrer' target='_blank' style={{ marginLeft: 8 }}>
+          <Link
+            href={item.ServerRedirectedURL}
+            rel='noopener noreferrer'
+            target='_blank'
+            style={{ marginLeft: 8 }}>
             {columnValue}
           </Link>
         </span>
