+++ SharePointFramework/PortfolioWebParts/src/components/PortfolioOverview/RenderItemColumn/TitleColumn/types.tsx
@@ -1,5 +1,5 @@
 import { IPortfolioOverviewProps } from 'components/PortfolioOverview'
-import { IFetchDataForViewItemResult } from 'data/IFetchDataForViewItemResult'
+import { IFetchDataForViewItemResult } from 'data/types'
 
 export interface ITitleColumnProps {
   props: IPortfolioOverviewProps
