--- SharePointFramework/PortfolioWebParts/src/data/IFetchDataForViewItemResult.ts
@@ -1,6 +1,0 @@
-import { SearchResult } from '@pnp/sp'
-
-export interface IFetchDataForViewItemResult extends SearchResult {
-  SiteId: string
-  [key: string]: any
-}
